Remove App namespace, move isInGlobalMode to root
