+++ packages/data-context/src/sources/index.ts
@@ -1,7 +1,6 @@
 /* eslint-disable padding-line-between-statements */
 // created by autobarrel, do not modify directly
 
-export * from './AppDataSource'
 export * from './BrowserDataSource'
 export * from './CloudDataSource'
 export * from './ConfigDataSource'
