+++ packages/frontend-shared/cypress/support/mock-graphql/stubgql-Browser.ts
@@ -1,7 +1,7 @@
 import type { Browser } from '../generated/test-graphql-types.gen'
 
 import { testNodeId } from './clientTestUtils'
-import { longBrowsersList } from './stubgql-App'
+import { longBrowsersList } from './longBrowsersList'
 
 export const stubBrowsers = longBrowsersList.map((browser, i): Browser => {
   return {
