+++ packages/frontend-shared/cypress/support/mock-graphql/stubgql-Query.ts
@@ -6,9 +6,6 @@ export const stubQuery: MaybeResolver<Query> = {
   dev () {
     return {}
   },
-  app (source, args, ctx) {
-    return ctx.app
-  },
   wizard (source, args, ctx) {
     return ctx.wizard
   },
@@ -20,5 +17,8 @@ export const stubQuery: MaybeResolver<Query> = {
   },
   isAuthBrowserOpened (source, args, ctx) {
     return ctx.isAuthBrowserOpened
+  },
+  isInGlobalMode (source, args, ctx) {
+    return !ctx.currentProject
   },
 }
