+++ packages/graphql/schemas/schema.graphql
@@ -2,12 +2,6 @@
 ### Do not make changes to this file directly
 
 
-"""Namespace for information related to the app"""
-type App {
-  """Whether the app is in global mode or not"""
-  isInGlobalMode: Boolean!
-}
-
 """Base error"""
 type BaseError {
   message: String
@@ -569,7 +563,6 @@ type ProjectPreferences {
 
 """The root "Query" type containing all entry fields for our querying"""
 type Query {
-  app: App!
   baseError: BaseError
 
   """An error related to finding a browser"""
@@ -601,6 +594,9 @@ type Query {
 
   """Whether the browser has been opened for auth or not"""
   isAuthBrowserOpened: Boolean!
+
+  """Whether the app is in global mode or not"""
+  isInGlobalMode: Boolean!
 
   """Metadata about the nagivation menu"""
   navigationMenu: NavigationMenu
