+++ packages/graphql/src/schemaTypes/objectTypes/gql-Query.ts
@@ -1,7 +1,6 @@
 import { objectType } from 'nexus'
 import { BaseError } from '.'
 import { ProjectLike } from '..'
-import { App } from './gql-App'
 import { CurrentProject } from './gql-CurrentProject'
 import { DevState } from './gql-DevState'
 import { NavigationMenu } from './gql-NavigationMenu'
@@ -14,11 +13,6 @@ export const Query = objectType({
     t.field('baseError', {
       type: BaseError,
       resolve: (root, args, ctx) => ctx.baseError,
-    })
-
-    t.nonNull.field('app', {
-      type: App,
-      resolve: (root, args, ctx) => ctx.appData,
     })
 
     t.field('navigationMenu', {
@@ -55,6 +49,11 @@ export const Query = objectType({
       resolve: (source, args, ctx) => {
         return ctx.wizardData.browserErrorMessage
       },
+    })
+
+    t.nonNull.boolean('isInGlobalMode', {
+      description: 'Whether the app is in global mode or not',
+      resolve: (source, args, ctx) => !ctx.currentProject,
     })
 
     t.nonNull.boolean('isAuthBrowserOpened', {
