+++ packages/graphql/src/schemaTypes/objectTypes/index.ts
@@ -1,7 +1,6 @@
 /* eslint-disable padding-line-between-statements */
 // created by autobarrel, do not modify directly
 
-export * from './gql-App'
 export * from './gql-BaseError'
 export * from './gql-Browser'
 export * from './gql-CurrentProject'
