+++ packages/launchpad/src/setup/OpenBrowserList.spec.tsx
@@ -1,6 +1,6 @@
 import { OpenBrowserListFragmentDoc } from '../generated/graphql-test'
 import OpenBrowserList from './OpenBrowserList.vue'
-import { longBrowsersList } from '@packages/frontend-shared/cypress/support/mock-graphql/stubgql-App'
+import { longBrowsersList } from '@packages/frontend-shared/cypress/support/mock-graphql/longBrowsersList'
 
 const launchButtonSelector = 'button[data-testid=launch-button]'
